dotenv in reqs.txt, moving utils up a level
