+++ server.py
@@ -1,17 +1,21 @@
-import base64
-import json
 import os
-
-from dotenv import load_dotenv
-from flask import Flask, render_template, request
+import json
 from worker import speech_to_text, text_to_speech, openai_process_message
+
+from flask import Flask, render_template, request
 from flask_cors import CORS
 
+from utils.logger import Logger
+from enums.logger import LogLevel
+
 # Load environment variables
+from dotenv import load_dotenv
 load_dotenv()
+
 
 app = Flask(__name__)
 cors = CORS(app, resources={r"/*": {"origins": "*"}})
+logger = Logger()
 
 @app.route('/', methods=['GET'])
 def index():
@@ -34,4 +38,9 @@ def process_prompt_route():
 
 
 if __name__ == "__main__":
+    # Sanity checking myself that we are loading
+    # env vars on a local machine
+    message = 'MY_ENV_VAR: ' + os.getenv('MY_ENV_VAR')
+    logger.log(LogLevel.INFO, message)
     app.run(port=8000, host='0.0.0.0')
+
